Refactored failure_zone test file and removed unnecessary tests
